CRediT Roles

CRediT Roles

Usage

credit_roles(display = c("explain", "names", "abbr"))

Arguments

display

Whether to display the category names, explanations, or abbreviations

Value

list of roles

Examples

credit_roles()
#> [1/con] Conceptualization: Ideas; formulation or evolution of overarching research goals and aims.
#> [2/dat] Data curation: Management activities to annotate (produce metadata), scrub data and maintain research data (including software code, where it is necessary for interpreting the data itself) for initial use and later re-use.
#> [3/ana] Formal analysis: Application of statistical, mathematical, computational, or other formal techniques to analyse or synthesize study data.
#> [4/fun] Funding acquisition: Acquisition of the financial support for the project leading to this publication.
#> [5/inv] Investigation: Conducting a research and investigation process, specifically performing the experiments, or data/evidence collection.
#> [6/met] Methodology: Development or design of methodology; creation of models.
#> [7/adm] Project administration: Management and coordination responsibility for the research activity planning and execution.
#> [8/res] Resources: Provision of study materials, reagents, materials, patients, laboratory samples, animals, instrumentation, computing resources, or other analysis tools.
#> [9/sof] Software: Programming, software development; designing computer programs; implementation of the computer code and supporting algorithms; testing of existing code components.
#> [10/sup] Supervision: Oversight and leadership responsibility for the research activity planning and execution, including mentorship external to the core team.
#> [11/val] Validation: Verification, whether as a part of the activity or separate, of the overall replication/reproducibility of results/experiments and other research outputs.
#> [12/vis] Visualization: Preparation, creation and/or presentation of the published work, specifically visualization/data presentation.
#> [13/dra] Writing - original draft: Preparation, creation and/or presentation of the published work, specifically writing the initial draft (including substantive translation).
#> [14/edi] Writing - review & editing: Preparation, creation and/or presentation of the published work by those from the original research group, specifically critical review, commentary or revision -- including pre- or post-publication stages.