List modules

List modules

Usage

module_list(module_dir = system.file("modules", package = "papercheck"))

Arguments

module_dir

the directory to search for modules (defaults to the built-in modules)

Value

a data frame of modules

Examples

module_list()
#>  * ai-summarise: Generate a 1-sentence summary for each section
#>  * all-p-values: List all p-values in the text, returning the matched text (e.g., 'p = 0.04') and document location in a table.
#>  * all-urls: List all the URLs in the main text
#>  * imprecise-p: List any p-values reported with insufficient precision (e.g., p < .05 or p = n.s.)
#>  * marginal: List all sentences that describe an effect as 'marginally significant'.
#>  * osf-check: List all OSF links and whether they are open, closed, or do not exist.
#>  * ref-consistency: Check if all references are cited and all citations are referenced
#>  * retractionwatch: Flag any cited papers in the RetractionWatch database
#>  * statcheck: Check consistency of p-values and test statistics