Run a module

Run a module

Usage

module_run(paper, module)

Arguments

paper

a paper object or a list of paper objects

module

the name of a module or path to a module to run on this object

Value

a list of the returned table and report text

Examples

filename <- demoxml()
paper <- read_grobid(filename)
module_run(paper, "imprecise-p")
#> ### Imprecise P-Values {.red}
#> 
#> You may have reported some imprecise p-values
#> 
#> |text    |section |header  | div|  p|  s|
#> |:-------|:-------|:-------|---:|--:|--:|
#> |p > .05 |results |Results |   3|  2|  2|