Convert a PDF to Grobid XML

This function uses the grobid server provided by Daniel Lakens at Technical University Eindhoven. You can set up your own local grobid server following instructions from https://grobid.readthedocs.io/ and set the argument grobid_url to the path to the processFullTextDocument API.

Usage

pdf2grobid(
  filename,
  save_path = ".",
  grobid_url = "https://grobid.work.abed.cloud/api/processFulltextDocument"
)

Arguments

filename

path to the PDF

save_path

directory or file path to save to; set to NULL to save to a temp file

grobid_url

the URL to the grobid server

Value

XML object