Retrieve info from ResearchBox by URL

Retrieve info from ResearchBox by URL

Usage

rbox_retrieve(rb_url, id_col = 1)

Arguments

rb_url

an ResearchBox URL, or a table containing them (e.g., as created by rbox_links())

id_col

the index or name of the column that contains ResearchBox URLs, if id is a table

Value

a data frame of information

Examples

# \donttest{
  # get info on one OSF node
  rbox_retrieve("https://researchbox.org/801")
#> Starting ResearchBox retrieval for 1 file...
#> * Retrieving info from https://researchbox.org/801...
#> ...ResearchBox retrieval complete!
#>                        rb_url
#> 1 https://researchbox.org/801
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   files
#> 1   READ ME.rtf,   Participant Survey_Redacted.pdf,   Informant Survey_Redacted.pdf,   Emotional_Intelligence_Survey_-_Subject_Pool_-_Participants.qsf,   Emotional_Intelligence_Survey_-_Subject_Pool_-_Informants.qsf,   Anonymized Data.csv,   Code for Analysis.R,   AsPredicted (for peer-review only) #50249.pdf,   Supplementary Materials 052923.docx,  Read-me file,  pdf - Qualtrics survey for focal participants,  pdf - Qualtrics survey for informants,  qsf - Qualtrics survey for focal participants,  qsf - Qualtrics survey for informants,  csv,  R Code,  Pre-registration,  Supplemental Materials
#>                                                                                                                                                                                                                                                                   RB_target
#> 1 He JC, Côté S. (2023) 'Are Empathic People Better Adjusted? A Test of Competing Models of Empathic Accuracy and Intrapersonal and Interpersonal Facets of Adjustment Using Self- and Peer Reports'.Psychological Science. V34(9):955-967.\ndoi: 10.1177/09567976231185127
#>                                                                                                                                                                                                                                                                                RB_license
#> 1 All content posted to ResearchBox is under a\r CC By 4.0 License(all use is allowed as long as authorship of the content is attributed). When using content from ResearchBox please\r cite the original work, and provide a link to the URL for this box (https://researchbox.org/801).
#>       RB_public
#> 1 June 09, 2023
#>                                                                                RB_authors
#> 1 Joyce He (joyce.he@anderson.ucla.edu)\nStéphane Côté (stephane.cote@rotman.utoronto.ca)
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                RB_abstract
#> 1 Are individuals adept at perceiving others’ emotions optimally adjusted? We extend past research by conducting a high-powered pre-registered study that comprehensively tests five theoretical models of empathic ability (i.e., emotion recognition ability) and self-views and intra- and interpersonal facets of adjustment in a sample of 1126 undergraduate students from Canada and 2205 informants. We obtained both self- and peer-reports of adjustment and controlled for cognitive abilities as a potential confounding variable. Empathic accuracy ability (but not self-views of that ability) was positively related to relationship satisfaction rated by both participants and informants. Self-views about empathic accuracy (but not actual empathic accuracy) were positively related to life satisfaction rated by both participants and informants. All associations held when controlling for cognitive abilities.
# }