Search the full text

Search the full text of a study paper. Currently only works with study objects that have full text imported from grobid (e.g., using study_from_xml()).

Usage

search_full_text(
  study,
  term,
  section = NULL,
  return = c("sentence", "paragraph", "section", "match"),
  refs = FALSE,
  ignore.case = TRUE,
  ...
)

Arguments

study

a study object created by study_from_xml or a list of study objects

term

the regex term to search for

section

the section(s) to search in

return

the kind of text to return, the full sentence, paragraph, or section that the text is in, or just the (regex) match

refs

whether to include references

ignore.case

whether to ignore case when text searching

...

additional arguments to pass to grepl()

Value

a data frame of matching sentences

Examples

grobid_dir <- system.file("grobid", package="scienceverse")
filename <- file.path(grobid_dir, "eyecolor.pdf.tei.xml")
study <- study_from_xml(filename)
#> Error in if (file.info(filename)$isdir) {    xmls <- list.files(filename, "\\.xml", full.names = TRUE)    if (length(xmls) == 0) {        stop("There are no xml files in the directory ", filename)    }    s <- study_from_xml(xmls)    return(s)}: missing value where TRUE/FALSE needed
sig <- search_full_text(study, "significant", "results")
#> Error in eval(expr, envir, enclos): object 'study' not found