Check Stats

Check Stats

Usage

stats(text, ...)

Arguments

text

the search table (or list of scienceverse objects)

...

arguments to pass to statcheck()

Value

a table of statistics

Examples

filename <- system.file("grobid", "incest.xml", package="papercheck")
study <- read_grobid(filename)
stats(study)
#>   test_type df1     df2 test_comp test_value p_comp reported_p  computed_p
#> 1         t  NA 578.003         =      2.359      =      0.019 0.018655985
#> 2         t  NA 578.000         =     -2.840      =      0.005 0.004670184
#> 3         t  NA 452.000         =      2.022      =      0.044 0.043764941
#> 4         t  NA 126.000         =     -1.933      =      0.055 0.055479644
#> 5         t  NA 578.000         =     -1.373      =      0.170 0.170284762
#>                             raw error decision_error one_tailed_in_txt
#> 1 t(578.003) = 2.359, p = 0.019 FALSE          FALSE             FALSE
#> 2     t(578) = -2.84, p = 0.005 FALSE          FALSE             FALSE
#> 3     t(452) = 2.022, p = 0.044 FALSE          FALSE             FALSE
#> 4    t(126) = -1.933, p = 0.055 FALSE          FALSE             FALSE
#> 5     t(578) = -1.373, p = 0.17 FALSE          FALSE             FALSE
#>   apa_factor
#> 1          1
#> 2          1
#> 3          1
#> 4          1
#> 5          1
#>                                                                                                                                                                                                                                                                                                              text
#> 1                                                                                                                                   Our main, omnibus analysis showed a significant three-way interaction among incest type, participant sex, and sibling type (Estimate = 0.285, t(578.003) = 2.359, p = 0.019).
#> 2                                                                                                                      Our analysis of moral wrongness scores for sibling incest scores showed a significant interaction between participant sex and sibling type (Estimate = -0.559, t(578) = -2.84, p = 0.005).
#> 3 Women with brothers only tended to view sibling incest more negatively than did women with sisters only (Estimate = 0.179, t(452) = 2.022, p = 0.044) while men with brothers only tended to view sibling incest less negatively than did men with sisters only (Estimate = -0.38, t(126) = -1.933, p = 0.055).
#> 4 Women with brothers only tended to view sibling incest more negatively than did women with sisters only (Estimate = 0.179, t(452) = 2.022, p = 0.044) while men with brothers only tended to view sibling incest less negatively than did men with sisters only (Estimate = -0.38, t(126) = -1.933, p = 0.055).
#> 5                                                                                                       In contrast, our analysis of moral wrongness scores for parental incest scores showed no significant interaction between participant sex and sibling type (Estimate = -0.268, t(578) = -1.373, p = 0.17).
#>   section  header div p s       file
#> 1 results Results   5 1 1 incest.xml
#> 2 results Results   5 2 1 incest.xml
#> 3 results Results   5 2 2 incest.xml
#> 4 results Results   5 2 2 incest.xml
#> 5 results Results   5 2 3 incest.xml