Check Stats

Check Stats

Usage

stats(text, ...)

Arguments

text

the search table (or list of scienceverse paper objects)

...

arguments to pass to statcheck()

Value

a table of statistics

Examples

dir <- demofile()
papers <- read_grobid(dir)
stats(papers)
#>    test_type df1     df2 test_comp test_value p_comp reported_p   computed_p
#> 1          Z  NA      NA         =      2.520      =      0.012 1.173548e-02
#> 2          Z  NA      NA         =     -0.190      =      0.849 8.493091e-01
#> 3          Z  NA      NA         =      2.355      =      0.019 1.852271e-02
#> 4          Z  NA      NA         =      0.152      =      0.879 8.791869e-01
#> 5          Z  NA      NA         =      1.113      =      0.266 2.657084e-01
#> 6          Z  NA      NA         =      0.836      =      0.403 4.031549e-01
#> 7          Z  NA      NA         =      1.515      =      0.130 1.297724e-01
#> 8          Z  NA      NA         =      0.960      =      0.337 3.370552e-01
#> 9          Z  NA      NA         =      1.645      =      0.100 9.996981e-02
#> 10         Z  NA      NA         =      2.188      =      0.029 2.866960e-02
#> 11         Z  NA      NA         =      2.207      =      0.027 2.731406e-02
#> 12         Z  NA      NA         =      1.774      =      0.076 7.606315e-02
#> 13         Z  NA      NA         =      2.188      =      0.091 2.866960e-02
#> 14         Z  NA      NA         =      1.292      =      0.196 1.963571e-01
#> 15         t  NA 578.003         =      2.359      =      0.019 1.865598e-02
#> 16         t  NA 578.000         =     -2.840      =      0.005 4.670184e-03
#> 17         t  NA 452.000         =      2.022      =      0.044 4.376494e-02
#> 18         t  NA 126.000         =     -1.933      =      0.055 5.547964e-02
#> 19         t  NA 578.000         =     -1.373      =      0.170 1.702848e-01
#> 20         t  NA  86.000         =     13.470      <      0.001 6.729803e-23
#> 21         t  NA  85.000         =     12.870      <      0.001 1.159479e-21
#> 22         t  NA 311.650         =     -3.450      <      0.001 6.378390e-04
#> 23         t  NA 288.610         =     -2.260      =      0.012 2.456660e-02
#> 24         t  NA 323.680         =     -4.550      <      0.001 7.597548e-06
#> 25         t  NA 305.340         =     -2.900      =      0.002 4.002409e-03
#>                              raw error decision_error one_tailed_in_txt
#> 1            z = 2.52, p = 0.012 FALSE          FALSE             FALSE
#> 2          z = --0.19, p = 0.849 FALSE          FALSE             FALSE
#> 3           z = 2.355, p = 0.019 FALSE          FALSE             FALSE
#> 4           z = 0.152, p = 0.879 FALSE          FALSE             FALSE
#> 5           z = 1.113, p = 0.266 FALSE          FALSE             FALSE
#> 6           z = 0.836, p = 0.403 FALSE          FALSE             FALSE
#> 7            z = 1.515, p = 0.13 FALSE          FALSE             FALSE
#> 8            z = 0.96, p = 0.337 FALSE          FALSE             FALSE
#> 9             z = 1.645, p = 0.1 FALSE          FALSE             FALSE
#> 10          Z = 2.188, p = 0.029 FALSE          FALSE             FALSE
#> 11          Z = 2.207, p = 0.027 FALSE          FALSE             FALSE
#> 12          Z = 1.774, p = 0.076 FALSE          FALSE             FALSE
#> 13          Z = 2.188, p = 0.091  TRUE           TRUE             FALSE
#> 14          Z = 1.292, p = 0.196 FALSE          FALSE             FALSE
#> 15 t(578.003) = 2.359, p = 0.019 FALSE          FALSE             FALSE
#> 16     t(578) = -2.84, p = 0.005 FALSE          FALSE             FALSE
#> 17     t(452) = 2.022, p = 0.044 FALSE          FALSE             FALSE
#> 18    t(126) = -1.933, p = 0.055 FALSE          FALSE             FALSE
#> 19     t(578) = -1.373, p = 0.17 FALSE          FALSE             FALSE
#> 20      t(86) = 13.47, p < 0.001 FALSE          FALSE             FALSE
#> 21      t(85) = 12.87, p < 0.001 FALSE          FALSE             FALSE
#> 22  t(311.65) = -3.45, p < 0.001 FALSE          FALSE             FALSE
#> 23  t(288.61) = -2.26, p = 0.012  TRUE          FALSE             FALSE
#> 24  t(323.68) = -4.55, p < 0.001 FALSE          FALSE             FALSE
#> 25  t(305.34) = -2.90, p = 0.002  TRUE          FALSE             FALSE
#>    apa_factor
#> 1           1
#> 2           1
#> 3           1
#> 4           1
#> 5           1
#> 6           1
#> 7           1
#> 8           1
#> 9           1
#> 10          1
#> 11          1
#> 12          1
#> 13          1
#> 14          1
#> 15          1
#> 16          1
#> 17          1
#> 18          1
#> 19          1
#> 20          1
#> 21          1
#> 22          1
#> 23          1
#> 24          1
#> 25          1
#>                                                                                                                                                                                                                                                                                                                                                                                                text
#> 1                                                                                                                                                                                                                                                                                                                           = 0.322, z = 2.52, p = 0.012), but not male partners (B = --0.059, S.E.
#> 2                                                                                                                                                                                                                                                                      = 0.31, z = --0.19, p = 0.849) and paternal eye color significantly predicted the eye color of male partners (B = 0.75, S.E.
#> 3                                                                                                                                                                                                                                                                                                                          = 0.318, z = 2.355, p = 0.019), but not female partners (B = 0.049, S.E.
#> 4                                                                                                                                                                                                                                                                                                                                                                    = 0.32, z = 0.152, p = 0.879).
#> 5                                                                                                                                                                                                                                                                                                                                                                   = 0.236, z = 1.113, p = 0.266).
#> 6                                                                                                                                                                                                                                                                                                                        = 0.313, z = 0.836, p = 0.403) more than women's partners (B = 0.472, S.E.
#> 7                                                                                                                                                                                                                                                                           = 0.311, z = 1.515, p = 0.13) and paternal eye color did not predict the eye color of women's partners (B = 0.309, S.E.
#> 8                                                                                                                                                                                                                                                                                                                           = 0.321, z = 0.96, p = 0.337) more than men's partners (B = 0.511, S.E.
#> 9                                                                                                                                                                                                                                                                                                                                                                     = 0.311, z = 1.645, p = 0.1).
#> 10                                                                                                                                                                                                                                                                 = 0.481, Z = 2.188, p = 0.029), while men's female partners' eye color was best predicted by maternal eye color (B = 1.072, S.E.
#> 11                                                                                                                                                                                                                                                                                                                                                                  = 0.486, Z = 2.207, p = 0.027).
#> 12                                                                                                                                                                                                                                                                                                                                                                  = 0.478, Z = 1.774, p = 0.076).
#> 13                                                                                                                                                                                                                                                           = 0.489, Z = 2.188, p = 0.091), maternal eye color was the only positive predictor in the first step of this analysis (B = 0.694, S.E.
#> 14                                                                                                                                                                                                                                                                                                                                                                  = 0.537, Z = 1.292, p = 0.196).
#> 15                                                                                                                                                                                                                    Our main, omnibus analysis showed a significant three-way interaction among incest type, participant sex, and sibling type (Estimate = 0.285, t(578.003) = 2.359, p = 0.019).
#> 16                                                                                                                                                                                                       Our analysis of moral wrongness scores for sibling incest scores showed a significant interaction between participant sex and sibling type (Estimate = -0.559, t(578) = -2.84, p = 0.005).
#> 17                                                                                  Women with brothers only tended to view sibling incest more negatively than did women with sisters only (Estimate = 0.179, t(452) = 2.022, p = 0.044) while men with brothers only tended to view sibling incest less negatively than did men with sisters only (Estimate = -0.38, t(126) = -1.933, p = 0.055).
#> 18                                                                                  Women with brothers only tended to view sibling incest more negatively than did women with sisters only (Estimate = 0.179, t(452) = 2.022, p = 0.044) while men with brothers only tended to view sibling incest less negatively than did men with sisters only (Estimate = -0.38, t(126) = -1.933, p = 0.055).
#> 19                                                                                                                                                                                        In contrast, our analysis of moral wrongness scores for parental incest scores showed no significant interaction between participant sex and sibling type (Estimate = -0.268, t(578) = -1.373, p = 0.17).
#> 20                                                                                                                                                                                                                                             The difference between measurements (M = 2.47, SD = 1.71) was large, t(86) = 13.47, p < 0.001, Hedges' g z = 1.43, 95% CI [1.13;1.73], see Figure 1.
#> 21                                                                                                                                                                                                                                The difference between measurements (M = 2.35, SD = 1.69) was analyzed with a dependent t-test, t(85) = 12.87, p < 0.001, Hedges' g z = 1.38, 95% CI [1.08;1.67].
#> 22 Four Welch's independent t-tests using the TOST approach indicated statistical equivalence for the difference between conditions for the perceived benefits, t(311.65) = -3.45, p < 0.001, costs, t(288.61) = -2.26, p = 0.012, how likely researchers were to pre-register, t(323.68) = -4.55, p < 0.001, and percentage of studies they planned to pre-register, t(305.34) = -2.90, p = 0.002.
#> 23 Four Welch's independent t-tests using the TOST approach indicated statistical equivalence for the difference between conditions for the perceived benefits, t(311.65) = -3.45, p < 0.001, costs, t(288.61) = -2.26, p = 0.012, how likely researchers were to pre-register, t(323.68) = -4.55, p < 0.001, and percentage of studies they planned to pre-register, t(305.34) = -2.90, p = 0.002.
#> 24 Four Welch's independent t-tests using the TOST approach indicated statistical equivalence for the difference between conditions for the perceived benefits, t(311.65) = -3.45, p < 0.001, costs, t(288.61) = -2.26, p = 0.012, how likely researchers were to pre-register, t(323.68) = -4.55, p < 0.001, and percentage of studies they planned to pre-register, t(305.34) = -2.90, p = 0.002.
#> 25 Four Welch's independent t-tests using the TOST approach indicated statistical equivalence for the difference between conditions for the perceived benefits, t(311.65) = -3.45, p < 0.001, costs, t(288.61) = -2.26, p = 0.012, how likely researchers were to pre-register, t(323.68) = -4.55, p < 0.001, and percentage of studies they planned to pre-register, t(305.34) = -2.90, p = 0.002.
#>    section
#> 1    intro
#> 2    intro
#> 3    intro
#> 4    intro
#> 5    intro
#> 6    intro
#> 7    intro
#> 8    intro
#> 9    intro
#> 10  method
#> 11  method
#> 12  method
#> 13  method
#> 14  method
#> 15 results
#> 16 results
#> 17 results
#> 18 results
#> 19 results
#> 20   intro
#> 21   intro
#> 22   intro
#> 23   intro
#> 24   intro
#> 25   intro
#>                                                                     header div
#> 1                                                                 [div-01]   1
#> 2                                                                 [div-01]   1
#> 3                                                                 [div-01]   1
#> 4                                                                 [div-01]   1
#> 5                                                                 [div-01]   1
#> 6                                                                 [div-01]   1
#> 7                                                                 [div-01]   1
#> 8                                                                 [div-01]   1
#> 9                                                                 [div-01]   1
#> 10                                                       Statistical tests   5
#> 11                                                       Statistical tests   5
#> 12                                                       Statistical tests   5
#> 13                                                       Statistical tests   5
#> 14                                                       Statistical tests   5
#> 15                                                                 Results   5
#> 16                                                                 Results   5
#> 17                                                                 Results   5
#> 18                                                                 Results   5
#> 19                                                                 Results   5
#> 20 Are One-tailed Tests and Sequential Analyses Appropriate in Psychology?   2
#> 21 Are One-tailed Tests and Sequential Analyses Appropriate in Psychology?   2
#> 22                                                                Attitude   3
#> 23                                                                Attitude   3
#> 24                                                                Attitude   3
#> 25                                                                Attitude   3
#>    p  s           id         file
#> 1  4  6 eyecolor.xml eyecolor.xml
#> 2  4  7 eyecolor.xml eyecolor.xml
#> 3  4  8 eyecolor.xml eyecolor.xml
#> 4  4  9 eyecolor.xml eyecolor.xml
#> 5  4 11 eyecolor.xml eyecolor.xml
#> 6  4 13 eyecolor.xml eyecolor.xml
#> 7  4 14 eyecolor.xml eyecolor.xml
#> 8  4 15 eyecolor.xml eyecolor.xml
#> 9  4 16 eyecolor.xml eyecolor.xml
#> 10 1 13 eyecolor.xml eyecolor.xml
#> 11 1 14 eyecolor.xml eyecolor.xml
#> 12 1 16 eyecolor.xml eyecolor.xml
#> 13 1 18 eyecolor.xml eyecolor.xml
#> 14 1 19 eyecolor.xml eyecolor.xml
#> 15 1  1   incest.xml   incest.xml
#> 16 2  1   incest.xml   incest.xml
#> 17 2  2   incest.xml   incest.xml
#> 18 2  2   incest.xml   incest.xml
#> 19 2  3   incest.xml   incest.xml
#> 20 8  8   prereg.xml   prereg.xml
#> 21 8 14   prereg.xml   prereg.xml
#> 22 5  4   prereg.xml   prereg.xml
#> 23 5  4   prereg.xml   prereg.xml
#> 24 5  4   prereg.xml   prereg.xml
#> 25 5  4   prereg.xml   prereg.xml